Add files via upload
Logistic Regression:
The logistic regression model achieved an overall accuracy of 73%, indicating the proportion of correctly predicted instances out of the total observations. However, it showed challenges in correctly identifying holidays, resulting in a relatively low recall of 60% for this class. The precision for holidays was 15%, suggesting that when the model predicted a day as a holiday, it was correct only 15% of the time. The weighted average F1-score, considering both precision and recall, was 79%.

Random Forest:
In contrast, the random forest classifier demonstrated exceptional performance, achieving a perfect accuracy of 100%. This indicates that the model correctly predicted all instances, both holidays and non-holidays, in the test set. The precision, recall, and F1-score for both classes (holidays and non-holidays) were all 100%, emphasizing the robustness and reliability of the random forest model for predicting days with and without holidays.

In summary, while the logistic regression model showed reasonable accuracy, the random forest model outperformed by providing flawless predictions, making it a highly reliable choice for the given task.